requirements.txt
torch>=1.7.0
numpy>=1.19.0
gymnasium>=0.28.0
pygame>=2.0.0
matplotlib>=3.3.0
pyyaml>=5.4.0

